fix: random

+++ random.h
@@ -1,6 +1,6 @@
 #ifndef RANDOM
 #define RANDOM
 
-unsigned long int getRandomNumber(unsigned long int n);
+int getRandomNumber(int n);
 
 #endif
